Added Second Drill
Added a second drill in the 5-10-5 drill and the script needed to run it. The script will allow more gamemodes to be easily made that use a set spawning of the cone. In this drill, players will move from left to right and back to left similar to the 5-10-5 drill used in sports.